+ added functionality for sorting one-line blocks
Until now, input of only the following format could be sorted:
>
>    selector(s) {
>        property: value;
>        property: value;
>        ...
>    }
>

Now, the input supports one-line ruleset blocks as well:
>
>    selector(s) { property: value; property: value; ... }
>

All non-empty lines that are not complete blocks, such as comments,
multiple selectors for the same ruleset block but placed on separate
lines, are removed. They do not appear in the sorted version. This
creates an issue, since only the last selector (if its line was a
complete ruleset block) is preserved in the sorted version.
sorted version.
Empty lines are preserved, because they could have been there for
better readability.

v0.2.0