+ fix: fixed Lax mode sorting issue casued by Strict mode continuity checker
Since code execution in Strict mode stops if an unacceptable line
is detected, a check needs to be performed for every line on input.
Lax mode, however, does not need this continuity check. Because of
this difference, an oversight caused sorting in Lax mode to break.

This has been fixed.

v0.6.0